Codigo refatorado.
