+++ README.md
@@ -1,8 +1,13 @@
 # Cores_ft
 
-Monitora a temperatua e frequências dos cores. 
+Monitora a temperatua e frequências dos cores.
 
 Para obter a frequencia dos cores foi usado o arquivo /proc/cpuinfo.
 
-Para obter as temperatura do cores foi utilizado o utilitario **sensors** do pacote **lm_sensors**
+Para obter as temperatura do cores foi utilizado o utilitáqrio **sensors** do pacote **lm_sensors**
 
+Para rodar basta fazer:
+
+```console
+python3 -m cores_ft
+```
