+++ cores_ft/__main__.py
@@ -1,0 +1,12 @@
+import os
+from .loop import loop_event
+
+
+def main():
+    try:
+        loop_event()
+    except KeyboardInterrupt:
+        os.system('clear')
+        print('Stop the monitoring.')
+
+main()
