+++ cores_ft/screen.py
@@ -1,0 +1,5 @@
+import os
+
+
+def clear_screen():
+    os.system('clear')
